Create crushing bug
Hello, I've tried to search for a book in my kindle voyage in greek and koreader crushed and shut down. I tried again in english and it worked just fine. Out of curiosity I searched again with the title in english but the selected language as greek and it still gave up results (in english though). Then I tried again the title in greek but with the english language selected and it crushed again. So something funky happens when my search includes greek characters.